+++ src/cwt/implementation/object.c
@@ -59,7 +59,7 @@ static uint32_t hash_string(const char* key, int length)
   uint32_t hash = 2166136261u;
   for (int i = 0 ; i < length; i++)
   {
-    hash ^= (uint8_t)key[i];
+    hash ^= (uint16_t)key[i];
     hash *= 16777619;
   }
   return hash;
